Fix division bug in int_vs_float.c

+++ C learning/int_vs_float.c
@@ -3,7 +3,8 @@
 int main() {
     int a = 5, b = 2;
 
-    printf("%d\n%f", a/b, a/b);
+    printf("%d\n%f", a/b, (float)a/b);
 
     return 0;
+    
 }
